Maj Block user V1
premier version sur ios

+++ FLOW/www/js/popup_option.js
@@ -3,15 +3,16 @@ var element_to_delete = {};
 
 function delete_flow_from_bdd(element) {
     element_to_copy = "flow_tittle";
-    $("#label_copy_button").text("copier le titre du flow");
-    $("#label_delete_button").text("supprimer le flow");
-    Popup("popup-option", true, 90);
+    $("#label_copy_button").text("Copier le titre du flow");
+    $("#label_report_button").text("Signaler le flow");
+    $("#label_delete_button").text("Supprimer le flow");
+    Popup("popup-option", true, 85);
 
     if (window.localStorage.getItem("user_private_id") == element.privateID) {
-        Popup("popup-option", true, 85);
+        Popup("popup-option", true, 80);
         element_to_delete.type = "flow";
         element_to_delete.element = element;
-        $("#delete_button").css("display", "block");
+        $("#delete_button").css("display", "table");
     }
 }
 
@@ -28,15 +29,16 @@ function delete_flow_from_html(element) {
 
 function delete_comment_from_bdd(element) {
     element_to_copy = "comment";
-    $("#label_copy_button").text("copier le commentaire");
-    $("#label_delete_button").text("supprimer le commentaire");
-    Popup("popup-option", true, 90);
+    $("#label_copy_button").text("Copier le commentaire");
+    $("#label_report_button").text("Signaler le commentaire");
+    $("#label_delete_button").text("Supprimer le commentaire");
+    Popup("popup-option", true, 85);
 
     if (window.localStorage.getItem("user_private_id") == element.private_Id) {
-        Popup("popup-option", true, 85);
+        Popup("popup-option", true, 80);
         element_to_delete.type = "comment";
         element_to_delete.element = element;
-        $("#delete_button").css("display", "block");
+        $("#delete_button").css("display", "table");
     }
 }
 
@@ -64,6 +66,38 @@ function delete_comment_from_html(element) {
     }
 }
 
+$("#report_button").on("touchend", function () {
+    if (element_to_copy == "flow_tittle") { // element_to_copy c'est juste l'elem selectionné
+        navigator.notification.confirm("Veux-tu vraiment signaler ce flow ?", function (id) {
+            if (id == 1) {
+                Popup("popup-option", false);
+                let data = {
+                    additionalData: {
+                        type: "report_flow",
+                        ObjectId: current_flow_block.ObjectId
+                    }
+                };
+                ServerManager.AddReportFlow(data);
+            }
+        }, "Confirmation", ["Oui", "Annuler"])
+
+    }
+    if (element_to_copy == "comment") { // element_to_copy c'est juste l'elem selectionné
+        navigator.notification.confirm("Veux-tu vraiment signaler ce commentaire ?", function (id) {
+            console.log(id);
+            if (id == 1) {
+                Popup("popup-option", false);
+                let data = {
+                    additionalData: {
+                        type: "report_comment"
+                    }
+                };
+                in_app_notif(data);
+            }
+        }, "Confirmation", ["Oui", "Annuler"])
+    }
+});
+
 $("#copy_button").on("touchend", function () {
     if (element_to_copy == "flow_tittle") {
         copyToClipboard($(current_flow_block.fpost_description).text());
@@ -78,17 +112,39 @@ $("#copy_button").on("touchend", function () {
 $("#delete_button").on("touchend", function () {
     if (element_to_delete) {
         if (element_to_delete.type == "flow") {
-            let data_delete_flow = {
-                //element: element,
-                ObjectId: element_to_delete.element.ObjectId
-            };
-            ServerManager.DeleteFlow(data_delete_flow, element_to_delete.element);
+            navigator.notification.confirm("Veux-tu vraiment supprimer ce flow ?", function (id) {
+                if (id == 1) {
+                    Popup("popup-option", false);
+                    let data = {
+                        additionalData: {
+                            type: "delete_flow"
+                        }
+                    };
+                    let data_delete_flow = {
+                        //element: element,
+                        ObjectId: element_to_delete.element.ObjectId
+                    };
+                    ServerManager.DeleteFlow(data_delete_flow, element_to_delete.element);
+                    in_app_notif(data);
+                }
+            }, "Confirmation", ["Oui", "Annuler"])
         } else {
-            let data_delete_comment = {
-                //element: element,
-                ObjectId: element_to_delete.element.Id
-            };
-            ServerManager.DeleteComment(data_delete_comment, element_to_delete.element);
+            navigator.notification.confirm("Veux-tu vraiment supprimer ce commentaire ?", function (id) {
+                if (id == 1) {
+                    Popup("popup-option", false);
+                    let data = {
+                        additionalData: {
+                            type: "delete_comment"
+                        }
+                    };
+                    let data_delete_comment = {
+                        //element: element,
+                        ObjectId: element_to_delete.element.Id
+                    };
+                    ServerManager.DeleteComment(data_delete_comment, element_to_delete.element);
+                    in_app_notif(data);
+                }
+            }, "Confirmation", ["Oui", "Annuler"])
         }
     } else {
         alert("Une erreur est survenue lors de la suppression de cet élément");
